+++ golem-worker-service-base/src/parser/expr/mod.rs
@@ -6,4 +6,5 @@ mod record;
 mod selection;
 mod sequence;
 mod tuple;
+mod constructor;
 mod util;
